Refine tests and add guard
