+++ Package.resolved
@@ -6,8 +6,8 @@
         "repositoryURL": "https://github.com/AudioKit/AudioKit",
         "state": {
           "branch": null,
-          "revision": "3360ddec30659447201a9ca4360d20cd924aab32",
-          "version": "5.3.0"
+          "revision": "e7f37266dc9200f04a2646b108acae98c87f8e46",
+          "version": "5.4.1"
         }
       }
     ]
